(feature) Ingest log lines in Combined Log Format
The agreement is that a file in combined log format[1] will be provided.

This commit deals with the parsing of a file in CLF into a temporary CSV file. This CSV file is an existing mechanism that handles the ingest process. It uses a 'hidden' table called `hits_staging` and various direct database transactions to provide a fast and robust ingest process. It is therefore worth the effort to add a parsing step to take advantage of the existing process.

Counting the log lines was an added complexity. The CSV ingest approach requires a grouped list of unique values for all other columns. This means we cannot multiple rows with the count of `1` as our approach:

```
host.a, 1, x
host.a, 1, x
host.a, 1, x
```

Another interesting decision was how to provide a host name for the ingest. Only destination IP addresses are provided by CLF, and only hostnames are allowed in the ingest. Since each IP address must be recognised as a `Host` later on in the process, we use the record to provide a matching hostname.

[1] http://fileformats.archiveteam.org/wiki/Combined_Log_Format