Fix Auth0 login
In c642ee50 we changed from using Zendesk (via Omniauth) as our auth
provider to using Auth0 (again, via Omniauth). James S tells me that he
was unable to test his changes locally at the time so we never confirmed
that the new login was actually working. I just tried using it now.

After clicking the 'Please Login' button, we redirected to GET
/auth/auth0, which gave a 404.

The Auth0 Rails Quickstart guide [1] suggests that this request needs to
be a POST:

> To prevent forged authentication requests, make sure that you add a link
> with a method of :post (as described below using the link_to function in
> Rails) or create a form with a CSRF token included.

I don't know if / why this was working with a GET when we used Zendesk
as the Auth0 provider, and I haven't investigated it.

I've changed it so that the 'Please Login' button makes a POST request
to /auth/auth0 directly. This removes the need for
AuthenticationController#new so I've got rid of that.

[1] https://auth0.com/docs/quickstart/webapp/rails#trigger-authentication