Revert "Only show organisations with sites"
This reverts commit f633fbd533a19b2b1c02a7bb0be4fc63f1377584.

I think that this tool was originally built to handle transitions for
hundreds of government departments, and so the developers wanted to
think of ways to show only the “relevant” organisations.

However, dxw’s use case is different. I think we will never have to deal
with such a large number of organisations within a single instance of
the Transition Tool. And in our experience of using the tool, it’s been
confusing to create a organisation, choose to not set up any sites yet,
and then not see it in the organisations list. There's no way for a
Transition Tool user to recover from this situation either - validations
prevent them from creating an organisation with the same slug again, but
there’s no way for them to delete the existing organisation using the
tool.